requirements.txt
flask==2.3.3
flask-mysqldb==1.0.1
werkzeug==2.3.7
tensorflow==2.13.0
opencv-python-headless==4.8.0.76
numpy==1.24.3
mediapipe==0.10.3
pyserial==3.5
requests==2.31.0